// api/consolidated/models/RemittanceScheduleOperationResult.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { EntityValidationError } from './EntityValidationError';
import type { RemittanceSchedule } from './RemittanceSchedule';

export type RemittanceScheduleOperationResult = {
    outcome?: RemittanceScheduleOperationResult.outcome;
    errors?: Array<EntityValidationError> | null;
    data?: RemittanceSchedule;
    correlationId?: string | null;
    status?: number;
    message?: string | null;
    trace?: string | null;
};

export namespace RemittanceScheduleOperationResult {

    export enum outcome {
        SUCCESS = 'Success',
        FAILURE = 'Failure',
    }


}